phone_number type changed from int to long

+++ src/com/javaparttwo/service/LoginService.java
@@ -57,7 +57,7 @@ public class LoginService {
 
 	    if (rs.next()) {
 		return new User(rs.getString("username"), null, rs.getString("first_name"), rs.getString("last_name"),
-			rs.getInt("phone_number"), rs.getString("email"), rs.getString("role_id"));
+			rs.getLong("phone_number"), rs.getString("email"), rs.getString("role_id"));
 	    }
 
 	} catch (SQLException e) {
